Effects building
I am working on the particle effects for torches and blood spatter. Basic character movement is done, but we need to work on effects, character and trap designs, and level designs. Basic music and sound effects are added here and can be manipulated and added onto. All folders are organized as well.